service-workers/service-worker/synced-state.https.html

<!doctype html>
<title>ServiceWorker: worker objects have synced state</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
// Tests that ServiceWorker objects representing the same Service Worker
// entity have the same state. JS-level equality is now required according to
// the spec.
'use strict';

function nextChange(worker) {
  return new Promise(function(resolve, reject) {
      worker.addEventListener('statechange', function handler(event) {
          try {
            worker.removeEventListener('statechange', handler);
            resolve(event.currentTarget.state);
          } catch (err) {
            reject(err);
          }
        });
    });
}

promise_test(async t => {
  const SCOPE = 'resources/synced-state';
  const SCRIPT = 'resources/empty-worker.js';
  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);

  t.add_cleanup(async() => {
    if (registration)
      await registration.unregister();
  })

  const worker = await registration.installing;
  let state = await nextChange(worker);
  assert_equals(state, 'installed', 'state - installed');
  assert_equals(registration.installing, null, 'registration.installing');
  assert_equals(registration.waiting, worker, 'registration.waiting');
  assert_equals(registration.waiting.state, 'installed',
                'registration.waiting.state');
  assert_equals(registration.active, null, 'registration.active');

  state = await nextChange(worker);
  assert_equals(state, 'activating', 'state - activating');
  assert_equals(registration.installing, null, 'registration.installing');
  assert_equals(registration.waiting, null, 'registration.waiting');
  assert_equals(registration.active, worker, 'registration.active');
  assert_equals(registration.active.state, 'activating',
                'registration.active.state');

  state = await nextChange(worker);
  assert_equals(state, 'activated', 'state - activated');
  assert_equals(registration.installing, null, 'registration.installing');
  assert_equals(registration.waiting, null, 'registration.waiting');
  assert_equals(registration.active, worker, 'registration.active');
  assert_equals(registration.active.state, 'activated',
                'registration.active.state');

}, 'worker objects for the same entity have the same state');
</script>
